client/pyproject.toml
[project]
name = "nsecchain"
version = "1.0.0"
description = "NSEC Cache Datastore Demo - Client Library"
requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.4.0",
    "click>=8.0.0",
    "pytest>=7.0.0",
]

[project.scripts]
prime = "scripts.prime:main"
verify = "scripts.verify_synthesis:main"
report = "scripts.report:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
